backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.20
transformers>=4.48.0
torch>=2.0.0
sentencepiece>=0.2.0
python-dotenv>=1.0.1
openai>=1.61.0
langchain>=0.3.17
opencc-python-reimplemented>=0.1.7
